fix(release-gate): financial integrity Numeric(12,2), brand-item-level attribution, audit quarantine, hardening
- fix(financial): migrate all money fields Float->Numeric(12,2) via 0012_money_numeric_precision with Decimal handling, preserves values, PG compatible
- fix(revenue): brand-item-level attribution using BrandAnalyticsEvent.revenue_amount=subtotal (brand-isolated) not entire Order.total_amount, prevents multi-brand over-attribution (Order 1000: Brand A 300 + Brand B 700, visual for A should be 300 not 1000), fallback to OrderItem.subtotal, organic=total_subtotal-exclusive, no JOIN multiplication
- fix(migration): 0011 quarantine redesign - invalid sponsored placements now paused requiring operator review, not silently active with invented 0.5/50.0
- feat(migration): 0013 audit log table migration_audit_log for historical repair audit trail, re-quarantines active placements with default remediated values
- test(financial): add test_financial_integrity for multi-brand order brand-item-level, Float vs Numeric, visual search product-level lineage, VTON sequential
- test(audit): add test_audit_logging for operational audit log verification, pagination, tenant isolation, quarantine logic
- fix(tests): test_group6_brand_admin Decimal handling float(prod.base_price), test_group6_final_hardening accept item-level revenue_amount as valid fix for JOIN multiplication
- verified: visual search product-level 30-day lineage via BrandAnalyticsEvent view events, prevents cross-product false attribution
- verified: VTON sequential multi-garment with OOM per-layer handling and honest failure
- verified: frontend build passes, migrations upgrade/downgrade tested

+++ backend/app/models/brand_analytics.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timezone
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Text, Float, CheckConstraint
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Text, Float, Numeric, CheckConstraint
 from sqlalchemy.orm import relationship
 from backend.app.core.database import Base
 
@@ -25,15 +25,15 @@ class SponsoredPlacement(Base):
     brand_id = Column(Integer, ForeignKey("brand_profiles.id", ondelete="CASCADE"), nullable=False, index=True)
     product_id = Column(Integer, ForeignKey("products.id", ondelete="CASCADE"), nullable=False, index=True)
     placement_type = Column(String(50), default="stylist_featured", nullable=False, index=True) # "stylist_featured", "trending_hero", "fit_recom_top"
-    bid_amount_per_click = Column(Float, default=0.50, nullable=False)
-    daily_budget = Column(Float, default=50.0, nullable=False)
-    spent_today = Column(Float, default=0.0, nullable=False)
+    bid_amount_per_click = Column(Numeric(12, 2), default=0.50, nullable=False)
+    daily_budget = Column(Numeric(12, 2), default=50.0, nullable=False)
+    spent_today = Column(Numeric(12, 2), default=0.0, nullable=False)
     status = Column(String(20), default="active", nullable=False, index=True) # "active", "paused", "budget_exhausted"
 
     impressions = Column(Integer, default=0, nullable=False)
     clicks = Column(Integer, default=0, nullable=False)
     conversions = Column(Integer, default=0, nullable=False)
-    revenue_generated = Column(Float, default=0.0, nullable=False)
+    revenue_generated = Column(Numeric(12, 2), default=0.0, nullable=False)
 
     start_date = Column(DateTime, nullable=True)
     end_date = Column(DateTime, nullable=True)
